add conditions in trigger

+++ mod-di-converter-storage-server/src/main/resources/templates/db_scripts/create_profile_associations.sql
@@ -36,10 +36,12 @@ CREATE OR REPLACE FUNCTION remove_related_wrappers_profile_associations()
 RETURNS trigger AS
 $$
 BEGIN
-   DELETE FROM ${myuniversity}_${mymodule}.profile_wrappers
-   USING  ${myuniversity}_${mymodule}.profile_associations
-   WHERE ${myuniversity}_${mymodule}.profile_wrappers.id = OLD.detail_wrapper_id;
-   RETURN OLD;
+    IF OLD.master_profile_type = 'ACTION_PROFILE' AND OLD.detail_profile_type IN ('ACTION_PROFILE', 'MAPPING_PROFILE', 'MATCH_PROFILE') OR
+           OLD.master_profile_type IN ('JOB_PROFILE', 'MATCH_PROFILE') AND OLD.detail_profile_type = 'MATCH_PROFILE' THEN
+            DELETE FROM ${myuniversity}_${mymodule}.profile_wrappers
+            WHERE id = OLD.detail_wrapper_id;
+        END IF;
+    RETURN OLD;
 END;
 $$ LANGUAGE plpgsql;
 
@@ -50,5 +52,5 @@ CREATE TRIGGER remove_related_wrappers_profile_associations
     AFTER DELETE
     ON ${myuniversity}_${mymodule}.profile_associations
     FOR EACH ROW
-    EXECUTE PROCEDURE ${myuniversity}_${mymodule}.remove_related_wrappers_profile_associations();
+    EXECUTE FUNCTION ${myuniversity}_${mymodule}.remove_related_wrappers_profile_associations();
